requirements.txt
pandas
chromadb
fastapi
uvicorn
python-dotenv
langgraph
langchain
langchain-google-genai
langchain-openai
langchain-community
langchain-ollama
langchain_chroma
langgraph-checkpoint-redis
